move normalize functions to functional.features.normalize
